+++ src/factory/bootstrap/factory/dashboard/__init__.py
@@ -1,0 +1,1 @@
+"""Dashboard hub RPC handler modules."""
